requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
celery==5.3.6
redis==5.0.1
docker==7.1.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
httpx==0.26.0
requests<2.32.0
PyYAML==6.0.1
